+++ sormas-ui/src/main/java/de/symeda/sormas/ui/configuration/DevModeView.java
@@ -102,10 +102,10 @@ import de.symeda.sormas.api.utils.fieldvisibility.FieldVisibilityCheckers;
 import de.symeda.sormas.api.visit.VisitDto;
 import de.symeda.sormas.api.visit.VisitStatus;
 import de.symeda.sormas.ui.UserProvider;
-import de.symeda.sormas.ui.configuration.dto.CaseGenerationConfig;
-import de.symeda.sormas.ui.configuration.dto.ContactGenerationConfig;
-import de.symeda.sormas.ui.configuration.dto.EventGenerationConfig;
-import de.symeda.sormas.ui.configuration.dto.SampleGenerationConfig;
+import de.symeda.sormas.ui.configuration.generate.CaseGenerationConfig;
+import de.symeda.sormas.ui.configuration.generate.ContactGenerationConfig;
+import de.symeda.sormas.ui.configuration.generate.EventGenerationConfig;
+import de.symeda.sormas.ui.configuration.generate.SampleGenerationConfig;
 import de.symeda.sormas.ui.configuration.validator.StringToNumberValidator;
 import de.symeda.sormas.ui.utils.ButtonHelper;
 import de.symeda.sormas.ui.utils.CssStyles;
@@ -768,7 +768,7 @@ public class DevModeView extends AbstractConfigurationView {
 		"Red blood cells in urine was indeterminate 2 weeks ago",
 		"-" };
 
-	private static void initializeRandomGenerator() {
+	private void initializeRandomGenerator() {
 		if (useManualSeed) {
 			randomGenerator = new Random(manualSeed);
 		} else {
@@ -1575,527 +1575,4 @@ public class DevModeView extends AbstractConfigurationView {
 	public void enter(ViewChangeEvent event) {
 		super.enter(event);
 	}
-
-//	private static class CaseGenerationConfig {
-//
-//		private String caseCount;
-//		private LocalDate startDate;
-//		private LocalDate endDate;
-//		private Disease disease;
-//		private RegionReferenceDto region;
-//		private DistrictReferenceDto district;
-//
-//		CaseGenerationConfig() {
-//			loadDefaultConfig();
-//		}
-//
-//		public void loadDefaultConfig() {
-//			caseCount = "10";
-//			startDate = LocalDate.now().minusDays(90);
-//			endDate = LocalDate.now();
-//			disease = null;
-//			region = null;
-//			district = null;
-//		}
-//
-//		public void loadPerformanceTestConfig() {
-//			caseCount = "50";
-//			startDate = LocalDate.now().minusDays(90);
-//			endDate = LocalDate.now();
-//			disease = Disease.CORONAVIRUS;
-//			region = null;
-//			district = null;
-//		}
-//
-//		public String getCaseCount() {
-//			return caseCount;
-//		}
-//
-//		public int getCaseCountAsNumber() {
-//			return StringUtils.isBlank(getCaseCount()) ? 0 : Integer.parseInt(getCaseCount());
-//		}
-//
-//		public void setCaseCount(String caseCount) {
-//			this.caseCount = caseCount;
-//		}
-//
-//		public LocalDate getStartDate() {
-//			return startDate;
-//		}
-//
-//		public void setStartDate(LocalDate startDate) {
-//			this.startDate = startDate;
-//		}
-//
-//		public LocalDate getEndDate() {
-//			return endDate;
-//		}
-//
-//		public void setEndDate(LocalDate endDate) {
-//			this.endDate = endDate;
-//		}
-//
-//		public Disease getDisease() {
-//			return disease;
-//		}
-//
-//		public void setDisease(Disease disease) {
-//			this.disease = disease;
-//		}
-//
-//		public RegionReferenceDto getRegion() {
-//			return region;
-//		}
-//
-//		public void setRegion(RegionReferenceDto region) {
-//			this.region = region;
-//		}
-//
-//		public DistrictReferenceDto getDistrict() {
-//			return district;
-//		}
-//
-//		public void setDistrict(DistrictReferenceDto district) {
-//			this.district = district;
-//		}
-//	}
-
-//	private static class ContactGenerationConfig {
-//
-//		private String contactCount;
-//		private LocalDate startDate;
-//		private LocalDate endDate;
-//		private Disease disease;
-//		private RegionReferenceDto region;
-//		private DistrictReferenceDto district;
-//		private boolean createWithoutSourceCases;
-//		private boolean createMultipleContactsPerPerson;
-//		private boolean createWithVisits;
-//
-//		ContactGenerationConfig() {
-//			loadDefaultConfig();
-//		}
-//
-//		public void loadDefaultConfig() {
-//			contactCount = "10";
-//			startDate = LocalDate.now().minusDays(90);
-//			endDate = LocalDate.now();
-//			disease = null;
-//			region = null;
-//			district = null;
-//			createWithoutSourceCases = false;
-//			createMultipleContactsPerPerson = false;
-//			createWithVisits = false;
-//		}
-//
-//		public void loadPerformanceTestConfig() {
-//			contactCount = "50";
-//			startDate = LocalDate.now().minusDays(90);
-//			endDate = LocalDate.now();
-//			disease = Disease.CORONAVIRUS;
-//			region = null;
-//			district = null;
-//			createWithoutSourceCases = false;
-//			createMultipleContactsPerPerson = false;
-//			createWithVisits = false;
-//		}
-//
-//		public String getContactCount() {
-//			return contactCount;
-//		}
-//
-//		public void setContactCount(String contactCount) {
-//			this.contactCount = contactCount;
-//		}
-//
-//		public int getContactCountAsNumber() {
-//			return Integer.parseInt(getContactCount());
-//		}
-//
-//		public LocalDate getStartDate() {
-//			return startDate;
-//		}
-//
-//		public void setStartDate(LocalDate startDate) {
-//			this.startDate = startDate;
-//		}
-//
-//		public LocalDate getEndDate() {
-//			return endDate;
-//		}
-//
-//		public void setEndDate(LocalDate endDate) {
-//			this.endDate = endDate;
-//		}
-//
-//		public Disease getDisease() {
-//			return disease;
-//		}
-//
-//		public void setDisease(Disease disease) {
-//			this.disease = disease;
-//		}
-//
-//		public RegionReferenceDto getRegion() {
-//			return region;
-//		}
-//
-//		public void setRegion(RegionReferenceDto region) {
-//			this.region = region;
-//		}
-//
-//		public DistrictReferenceDto getDistrict() {
-//			return district;
-//		}
-//
-//		public void setDistrict(DistrictReferenceDto district) {
-//			this.district = district;
-//		}
-//
-//		public boolean isCreateWithoutSourceCases() {
-//			return createWithoutSourceCases;
-//		}
-//
-//		public void setCreateWithoutSourceCases(boolean createWithoutSourceCases) {
-//			this.createWithoutSourceCases = createWithoutSourceCases;
-//		}
-//
-//		public boolean isCreateMultipleContactsPerPerson() {
-//			return createMultipleContactsPerPerson;
-//		}
-//
-//		public void setCreateMultipleContactsPerPerson(boolean createMultipleContactsPerPerson) {
-//			this.createMultipleContactsPerPerson = createMultipleContactsPerPerson;
-//		}
-//
-//		public boolean isCreateWithVisits() {
-//			return createWithVisits;
-//		}
-//
-//		public void setCreateWithVisits(boolean createWithVisits) {
-//			this.createWithVisits = createWithVisits;
-//		}
-//	}
-
-//	private static class EventGenerationConfig {
-//
-//		private int eventCount;
-//		private LocalDate startDate;
-//		private LocalDate endDate;
-//		private Disease disease;
-//		private RegionReferenceDto region;
-//		private DistrictReferenceDto district;
-//		private int minParticipantsPerEvent;
-//		private int maxParticipantsPerEvent;
-//		private int minContactsPerParticipant;
-//		private int maxContactsPerParticipant;
-//		private int percentageOfCases;
-//
-//		EventGenerationConfig() {
-//			loadDefaultConfig();
-//		}
-//
-//		public void loadDefaultConfig() {
-//			eventCount = 10;
-//			startDate = LocalDate.now().minusDays(90);
-//			endDate = LocalDate.now();
-//			disease = null;
-//			region = null;
-//			district = null;
-//			minParticipantsPerEvent = 3;
-//			maxParticipantsPerEvent = 10;
-//			minContactsPerParticipant = 0;
-//			maxContactsPerParticipant = 3;
-//			percentageOfCases = 20;
-//		}
-//
-//		public void loadPerformanceTestConfig() {
-//			eventCount = 30;
-//			startDate = LocalDate.now().minusDays(90);
-//			endDate = LocalDate.now();
-//			disease = Disease.CORONAVIRUS;
-//			// region?
-//			// district?
-//			minParticipantsPerEvent = 3;
-//			maxParticipantsPerEvent = 8;
-//			minContactsPerParticipant = 0;
-//			maxContactsPerParticipant = 2;
-//			percentageOfCases = 15;
-//		}
-//
-//		public int getEventCount() {
-//			return eventCount;
-//		}
-//
-//		public void setEventCount(int eventCount) {
-//			this.eventCount = eventCount;
-//		}
-//
-//		public LocalDate getStartDate() {
-//			return startDate;
-//		}
-//
-//		public void setStartDate(LocalDate startDate) {
-//			this.startDate = startDate;
-//		}
-//
-//		public LocalDate getEndDate() {
-//			return endDate;
-//		}
-//
-//		public void setEndDate(LocalDate endDate) {
-//			this.endDate = endDate;
-//		}
-//
-//		public Disease getDisease() {
-//			return disease;
-//		}
-//
-//		public void setDisease(Disease disease) {
-//			this.disease = disease;
-//		}
-//
-//		public RegionReferenceDto getRegion() {
-//			return region;
-//		}
-//
-//		public void setRegion(RegionReferenceDto region) {
-//			this.region = region;
-//		}
-//
-//		public DistrictReferenceDto getDistrict() {
-//			return district;
-//		}
-//
-//		public void setDistrict(DistrictReferenceDto district) {
-//			this.district = district;
-//		}
-//
-//		public int getMinParticipantsPerEvent() {
-//			return minParticipantsPerEvent;
-//		}
-//
-//		public void setMinParticipantsPerEvent(int minParticipantsPerEvent) {
-//			this.minParticipantsPerEvent = minParticipantsPerEvent;
-//		}
-//
-//		public int getMaxParticipantsPerEvent() {
-//			return maxParticipantsPerEvent;
-//		}
-//
-//		public void setMaxParticipantsPerEvent(int maxParticipantsPerEvent) {
-//			this.maxParticipantsPerEvent = maxParticipantsPerEvent;
-//		}
-//
-//		public int getMinContactsPerParticipant() {
-//			return minContactsPerParticipant;
-//		}
-//
-//		public void setMinContactsPerParticipant(int minContactsPerParticipant) {
-//			this.minContactsPerParticipant = minContactsPerParticipant;
-//		}
-//
-//		public int getMaxContactsPerParticipant() {
-//			return maxContactsPerParticipant;
-//		}
-//
-//		public void setMaxContactsPerParticipant(int maxContactsPerParticipant) {
-//			this.maxContactsPerParticipant = maxContactsPerParticipant;
-//		}
-//
-//		public int getPercentageOfCases() {
-//			return percentageOfCases;
-//		}
-//
-//		public void setPercentageOfCases(int percentageOfCases) {
-//			this.percentageOfCases = percentageOfCases;
-//			if (this.percentageOfCases >= 100) {
-//				this.percentageOfCases = 100;
-//			} else if (this.percentageOfCases <= 0) {
-//				this.percentageOfCases = 0;
-//			}
-//		}
-//
-//	}
-
-//	private static class SampleGenerationConfig {
-//
-//		private int sampleCount;
-//		private SamplePurpose samplePurpose;
-//		private LocalDate startDate;
-//		private LocalDate endDate;
-//		private SampleMaterial sampleMaterial;
-//		private String sampleMaterialText;
-//		private FacilityReferenceDto laboratory;
-//
-//		private boolean externalLabOrInternalInHouseTesting = false;
-//		private boolean requestPathogenTestsToBePerformed = false;
-//		private boolean requestAdditionalTestsToBePerformed = false;
-//		private boolean sendDispatch = false;
-//		private boolean received = false;
-//		private String comment;
-//
-//		private Disease disease;
-//		private RegionReferenceDto region;
-//		private DistrictReferenceDto district;
-//
-//		private SampleGenerationConfig() {
-//		}
-//
-//		public static SampleGenerationConfig getDefaultConfig() {
-//			SampleGenerationConfig sampleGenerationConfig = new SampleGenerationConfig();
-//			sampleGenerationConfig.sampleCount = 10;
-//			sampleGenerationConfig.startDate = LocalDate.now().minusDays(90);
-//			sampleGenerationConfig.endDate = LocalDate.now();
-//			sampleGenerationConfig.disease = null;
-//			sampleGenerationConfig.region = null;
-//			sampleGenerationConfig.district = null;
-//			sampleGenerationConfig.samplePurpose = SamplePurpose.INTERNAL;
-//			sampleGenerationConfig.sampleMaterial = SampleMaterial.BLOOD;
-//			return sampleGenerationConfig;
-//		}
-//
-//		public static SampleGenerationConfig getPerformanceTestConfig() {
-//			SampleGenerationConfig sampleGenerationConfig = new SampleGenerationConfig();
-//			sampleGenerationConfig.sampleCount = 50;
-//			sampleGenerationConfig.startDate = LocalDate.now().minusDays(90);
-//			sampleGenerationConfig.endDate = LocalDate.now();
-//			sampleGenerationConfig.disease = Disease.CORONAVIRUS;
-//			sampleGenerationConfig.region = null;
-//			sampleGenerationConfig.district = null;
-//			sampleGenerationConfig.samplePurpose = SamplePurpose.EXTERNAL;
-//			sampleGenerationConfig.sampleMaterial = SampleMaterial.BLOOD;
-//			sampleGenerationConfig.laboratory = FacadeProvider.getFacilityFacade().getAllActiveLaboratories(false).get(0);
-//			return sampleGenerationConfig;
-//		}
-//
-//		public SamplePurpose getSamplePurpose() {
-//			return samplePurpose;
-//		}
-//
-//		public void setSamplePurpose(SamplePurpose samplePurpose) {
-//			this.samplePurpose = samplePurpose;
-//		}
-//
-//		public SampleMaterial getSampleMaterial() {
-//			return sampleMaterial;
-//		}
-//
-//		public void setSampleMaterial(SampleMaterial sampleMaterial) {
-//			this.sampleMaterial = sampleMaterial;
-//		}
-//
-//		public String getSampleMaterialText() {
-//			return sampleMaterialText;
-//		}
-//
-//		public void setSampleMaterialText(String sampleMaterialText) {
-//			this.sampleMaterialText = sampleMaterialText;
-//		}
-//
-//		public FacilityReferenceDto getLaboratory() {
-//			return laboratory;
-//		}
-//
-//		public void setLaboratory(FacilityReferenceDto laboratory) {
-//			this.laboratory = laboratory;
-//		}
-//
-//		public int getSampleCount() {
-//			return sampleCount;
-//		}
-//
-//		public void setSampleCount(int contactCount) {
-//			this.sampleCount = contactCount;
-//		}
-//
-//		public LocalDate getStartDate() {
-//			return startDate;
-//		}
-//
-//		public void setStartDate(LocalDate startDate) {
-//			this.startDate = startDate;
-//		}
-//
-//		public LocalDate getEndDate() {
-//			return endDate;
-//		}
-//
-//		public void setEndDate(LocalDate endDate) {
-//			this.endDate = endDate;
-//		}
-//
-//		public boolean isRequestPathogenTestsToBePerformed() {
-//			return requestPathogenTestsToBePerformed;
-//		}
-//
-//		public void setRequestPathogenTestsToBePerformed(boolean requestPathogenTestsToBePerformed) {
-//			this.requestPathogenTestsToBePerformed = requestPathogenTestsToBePerformed;
-//		}
-//
-//		public boolean isExternalLabOrInternalInHouseTesting() {
-//			return externalLabOrInternalInHouseTesting;
-//		}
-//
-//		public boolean isRequestAdditionalTestsToBePerformed() {
-//			return requestAdditionalTestsToBePerformed;
-//		}
-//
-//		public void setRequestAdditionalTestsToBePerformed(boolean requestAdditionalTestsToBePerformed) {
-//			this.requestAdditionalTestsToBePerformed = requestAdditionalTestsToBePerformed;
-//		}
-//
-//		public void setExternalLabOrInternalInHouseTesting(boolean externalLabOrInternalInHouseTesting) {
-//			this.externalLabOrInternalInHouseTesting = externalLabOrInternalInHouseTesting;
-//		}
-//
-//		public boolean isSendDispatch() {
-//			return sendDispatch;
-//		}
-//
-//		public void setSendDispatch(boolean sendDispatch) {
-//			this.sendDispatch = sendDispatch;
-//		}
-//
-//		public boolean isReceived() {
-//			return received;
-//		}
-//
-//		public void setReceived(boolean received) {
-//			this.received = received;
-//		}
-//
-//		public String getComment() {
-//			return comment;
-//		}
-//
-//		public void setComment(String comment) {
-//			this.comment = comment;
-//		}
-//
-//		public Disease getDisease() {
-//			return disease;
-//		}
-//
-//		public void setDisease(Disease disease) {
-//			this.disease = disease;
-//		}
-//
-//		public RegionReferenceDto getRegion() {
-//			return region;
-//		}
-//
-//		public void setRegion(RegionReferenceDto region) {
-//			this.region = region;
-//		}
-//
-//		public DistrictReferenceDto getDistrict() {
-//			return district;
-//		}
-//
-//		public void setDistrict(DistrictReferenceDto district) {
-//			this.district = district;
-//		}
-//
-//	}
 }
